chore: update certificate and packager script
Updated the certificate to the now required RSA-2048 and extend the
expiration date. Modified Packager.bat to provide the certificate
password directly.

Note that signing an AIR app is required to package it into a
captive runtime. For ease of packaging, Swivel uses a self-signed
certificate and is storing the certificate with password directly
in the repo. This certficiate is simply used to package the app,
and no claim is made of its security.