chore: look for AIR_SDK environemnt var
SetupSDK.bat will now look for the AIR_SDK environment var to
specify the path to the Adobe AIR SDK. If this var is unset, the
path will be set to location that HaxeDevelop installs to by
default. The AIR SDK can be installed from HaxeDevelop by going to
Tools->Install Software and checking AIR SDK + ASC 2.0.

This will pave the way to automated packaging via AppVeyor and
Travis.